refactore(helpers): add declineReviewWord utility and move mocks to separate file
