+++ components/mockProductCard/MockProductCard.interface.ts
@@ -1,4 +1,4 @@
-import { MockProduct } from '@/helpers';
+import { MockProduct } from '@/mocks/demoProducts.mock';
 
 export interface MockProductCardProps {
   product: MockProduct;
